% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.universe.api-package.R
\docType{package}
\name{r.universe.api-package}
\alias{r.universe.api}
\alias{r.universe.api-package}
\title{r.universe.api: Client for R-universe APIs}
\description{
Client for R-universe APIs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/r.universe.api}
  \item Report bugs at \url{https://github.com/vgherard/r.universe.api/issues}
}

}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
