% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runi-package.R
\docType{package}
\name{runi-package}
\alias{runi}
\alias{runi-package}
\title{runi: Client for R-universe APIs}
\description{
Client for R-universe APIs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/runi}
  \item Report bugs at \url{https://github.com/vgherard/runi/issues}
}

}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
