% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runi_badges.R
\name{runi_badges}
\alias{runi_badges}
\title{R-universe badges}
\usage{
runi_badges(universe, package)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}

\item{package}{either \code{NULL}, or a length one character. Package name
(see details).}
}
\value{
a character vector.
}
\description{
Returns the URL of the badge with current version of given
package.
}
\examples{
runi_badges("vgherard", "runi")

}
\author{
Valerio Gherardi
}
