% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runi_repos.R
\name{runi_repos}
\alias{runi_repos}
\title{R-universe repository base URL}
\usage{
runi_repos(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a string.
}
\description{
Get the base URL of an R-universe repository
}
\author{
Valerio Gherardi
}
