% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runi_stats_descriptions.R
\name{runi_stats_descriptions}
\alias{runi_stats_descriptions}
\title{DESCRIPTIONs}
\usage{
runi_stats_descriptions(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a \code{\link[tibble]{tibble}}.
}
\description{
Get DESCRIPTION files of packages available in an R-universe.
}
\examples{
runi_stats_descriptions("vgherard")

}
\author{
Valerio Gherardi
}
