% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runi_stats_sysdeps.R
\name{runi_stats_sysdeps}
\alias{runi_stats_sysdeps}
\title{External Dependencies}
\usage{
runi_stats_sysdeps(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a \code{\link[tibble]{tibble}}.
}
\description{
Get list with external libraries used in an R-universe universe.
}
\examples{
runi_stats_sysdeps("vgherard")

}
\author{
Valerio Gherardi
}
