% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runi_universes.R
\name{runi_universes}
\alias{runi_universes}
\title{List R-universes}
\usage{
runi_universes()
}
\value{
A tibble.
}
\description{
List all available R-universes.
}
\examples{
runi_universes()

}
\author{
Valerio Gherardi
}
