% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv-package.R
\docType{package}
\name{runiv-package}
\alias{runiv}
\alias{runiv-package}
\title{runiv: Client for R-universe APIs}
\description{
Client for R-universe APIs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/runiv}
  \item Report bugs at \url{https://github.com/vgherard/runiv/issues}
}

}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
