% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_cran_like.R
\name{runiv_cran_like}
\alias{runiv_cran_like}
\alias{runiv_src}
\alias{runiv_bin}
\title{CRAN-like Repository Structure}
\usage{
runiv_src(universe)

runiv_bin(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a \code{data.frame}.
}
\description{
CRAN-like repository structure for packages in an R-universe.
}
\examples{
runiv_src("vgherard")
}
\author{
Valerio Gherardi
}
