% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_descriptions.R
\name{runiv_descriptions}
\alias{runiv_descriptions}
\title{DESCRIPTIONs in an R-universe}
\usage{
runiv_descriptions(owner)
}
\arguments{
\item{owner}{A length one character. Name of user or organization
owner of the R-universe repository.}
}
\value{
a \code{data.frame}.
}
\description{
Get DESCRIPTIONs of packages available in an R-universe.
}
\examples{
runiv_descriptions("vgherard")

}
