% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_maintainers.R
\name{runiv_maintainers}
\alias{runiv_maintainers}
\title{List maintainers of an R-universe}
\usage{
runiv_maintainers(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a \code{data.frame}.
}
\description{
Lists all maintainers, identified by email address,
of an R-universe.
}
\examples{
runiv_maintainers("vgherard")

}
\author{
Valerio Gherardi
}
