% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_repos.R
\name{runiv_repos}
\alias{runiv_repos}
\title{R-universe repository base URL}
\usage{
runiv_repos(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a string.
}
\description{
Get the base URL of an R-universe repository
}
\author{
Valerio Gherardi
}
