% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_stats_checks.R
\name{runiv_stats_checks}
\alias{runiv_stats_checks}
\title{Recent Builds and Checks}
\usage{
runiv_stats_checks(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a \code{\link[tibble]{tibble}}.
}
\description{
Get list with recent builds and checks of packages available in
an R-universe.
}
\examples{
runiv_stats_checks("vgherard")

}
\author{
Valerio Gherardi
}
