% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_stats_descriptions.R
\name{runiv_stats_descriptions}
\alias{runiv_stats_descriptions}
\title{DESCRIPTIONs}
\usage{
runiv_stats_descriptions(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a \code{data.frame}.
}
\description{
Get DESCRIPTION files of packages available in an R-universe.
}
\examples{
runiv_stats_descriptions("vgherard")

}
\author{
Valerio Gherardi
}
