% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_stats_maintainers.R
\name{runiv_stats_maintainers}
\alias{runiv_stats_maintainers}
\title{List maintainers of an R-universe}
\usage{
runiv_stats_maintainers(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a \code{\link[tibble]{tibble}}.
}
\description{
Lists all maintainers, identified by email address,
of an R-universe.
}
\examples{
runiv_stats_maintainers("vgherard")

}
\author{
Valerio Gherardi
}
