% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_stats_sysdeps.R
\name{runiv_stats_sysdeps}
\alias{runiv_stats_sysdeps}
\title{External Dependencies}
\usage{
runiv_stats_sysdeps(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a \code{data.frame}.
}
\description{
Get list with external libraries used in an R-universe universe.
}
\examples{
runiv_stats_sysdeps("vgherard")

}
\author{
Valerio Gherardi
}
