% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_universes.R
\name{runiv_universes}
\alias{runiv_universes}
\title{List R-universes}
\usage{
runiv_universes()
}
\value{
A character vector
}
\description{
List all available R-universes.
}
\examples{
runiv_universes()

}
\author{
Valerio Gherardi
}
