% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_descriptions.R
\name{runiv_descriptions}
\alias{runiv_descriptions}
\title{DESCRIPTIONs in an R-universe}
\usage{
runiv_descriptions(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a \code{data.frame}.
}
\description{
Get DESCRIPTIONs of packages available in an R-universe.
}
\examples{
runiv_descriptions("vgherard")

}
\author{
Valerio Gherardi
}
