% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_packages.R
\name{runiv_packages}
\alias{runiv_packages}
\title{Packages in an R-universe}
\usage{
runiv_packages(universe)
}
\arguments{
\item{universe}{A length one character. User or organization name of the
the R-universe owner.}
}
\value{
a character vector. List of R packages in the \code{owner}'s
R-universe.
}
\description{
List all packages available in an R-universe.
}
\examples{
runiv_packages("vgherard")

}
\author{
Valerio Gherardi
}
