% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runiv_packages.R
\name{runiv_packages}
\alias{runiv_packages}
\title{R-universe API request}
\usage{
runiv_packages(owner)
}
\arguments{
\item{owner}{A length one character. Name of user or organization
owner of the R-universe repository.}
}
\value{
a character vector. List of R packages in the \code{owner}'s
R-universe.
}
\description{
Make API request to an R-universe.
}
\examples{
runiv_packages("vgherard")

}
